Fixed padding issue in TextInput (#1630)
ReactTextInputShadow uses a measure function to compute text size. We also used to piggyback on that callback to retrieve the computed padding (that we do push to the TextBox controls).

The problem is that Yoga refrains from calling the measuring function when it already knows the sizes (usually due to a parent enforcing a particular size irrespective of text content).

I moved the retrieval of computed padding to the "collection of extra data" point.